node/gympass-solid/src/use-cases/fetch-user-check-ins-history.spec.ts
import { beforeEach, describe, expect, it } from "vitest";

import { InMemoryCheckInRepository } from "@/repositories/in-memory-repositories/in-memory-checkin-repository";

import { FetchUserCheckInsHistoryUseCase } from "./fetch-user-check-ins-history";

let checkInRepository: InMemoryCheckInRepository;
// SUT - System Under Teste (Sistema em teste) trata-se do ator que está sendo testado nesse caso a classe instanciada abaixo
let sut: FetchUserCheckInsHistoryUseCase;

describe("Fetch check-in history Use Case", () => {
  beforeEach(async () => {
    checkInRepository = new InMemoryCheckInRepository();
    sut = new FetchUserCheckInsHistoryUseCase(checkInRepository);
  });

  it("should be able to fetch user's history check ins", async () => {
    await checkInRepository.create({
      gym_id: "gym-01",
      user_id: "user-01",
    });
    await checkInRepository.create({
      gym_id: "gym-02",
      user_id: "user-01",
    });

    const { checkIns } = await sut.execute({
      userId: "user-01",
      page: 1,
    });

    expect(checkIns.length).toBe(2);
    expect(checkIns).toEqual([
      expect.objectContaining({ gym_id: "gym-01" }),
      expect.objectContaining({ gym_id: "gym-02" }),
    ]);
  });

  it("should be able to fetch paginated user's history check ins", async () => {
    await Promise.all(
      Array.from({ length: 22 }).map((_, i) =>
        checkInRepository.create({
          gym_id: `gym-${i + 1}`,
          user_id: "user-01",
        })
      )
    );
    const { checkIns } = await sut.execute({
      userId: "user-01",
      page: 2,
    });

    expect(checkIns.length).toBe(2);
    expect(checkIns).toEqual([
      expect.objectContaining({ gym_id: "gym-21" }),
      expect.objectContaining({ gym_id: "gym-22" }),
    ]);
  });
});
